Rename: 파일 위치를 constants/로 이동
- 상수 파일이므로 파일 위치를 constants/로 이동
